Remove `src` from `files`
`files` specifies what has to be published to npm.
We don't need to publish source code, as we should publish only the
results of the build in `dist` and `declarations`.

Move AST transform up and publish lib folder as well.

No need to keep .npmignore as `files` whitelist the files.